ApiErrorResponse - PortfolioPosts - Cloudinary - RoomImage and descri…

+++ Roomify.GP.API/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Roomify.GP.API.Errors;
 using Roomify.GP.Core.DTOs.User;
 using Roomify.GP.Core.Services.Contract;
 using System;
@@ -7,7 +8,7 @@ using System.Threading.Tasks;
 namespace Roomify.GP.API.Controllers
 {
     [ApiController]
-    [Route("api/users")]
+    [Route("api/[controller]")]
     public class UsersController : ControllerBase
     {
         private readonly IUserService _userService;
@@ -17,13 +18,9 @@ namespace Roomify.GP.API.Controllers
             _userService = userService;
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetUser(Guid id)
-        {
-            var user = await _userService.GetUserByIdAsync(id);
-            return user != null ? Ok(user) : NotFound();
-        }
 
+
+        [ProducesResponseType(typeof(IActionResult), StatusCodes.Status200OK)]
         [HttpGet]
         public async Task<IActionResult> GetAllUsers()
         {
@@ -31,18 +28,36 @@ namespace Roomify.GP.API.Controllers
             return Ok(users);
         }
 
+
+        [ProducesResponseType(typeof(IActionResult), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
+        [HttpGet("{id}")]
+        public async Task<IActionResult> GetUser(Guid id)
+        {
+            var user = await _userService.GetUserByIdAsync(id);
+            return user != null ? Ok(user) : NotFound(new ApiErrorResponse(404));
+        }
+
+
+        [ProducesResponseType(typeof(IActionResult), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateUser(Guid id, UserUpdateDto userDto)
         {
             var result = await _userService.UpdateUserAsync(id, userDto);
-            return result ? NoContent() : NotFound();
+            return result ? Ok(userDto) : NotFound(new ApiErrorResponse(404));
         }
 
+
+        [ProducesResponseType(typeof(IActionResult), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status404NotFound)]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUser(Guid id)
         {
+            if (id == null) return BadRequest(new ApiErrorResponse(400));
             var result = await _userService.DeleteUserAsync(id);
-            return result ? NoContent() : NotFound();
+            return result ? Ok("User Deleted Successfully") : NotFound(new ApiErrorResponse(404));
         }
     }
 }
